feat: migrate `/api/users` endpoint to `/api/v1/users` (#142)
# Description

This PR add a new endpoint `GET /api/v1/users` returning a list of users
if the current user is an owner.

Changes of the new endpoint compared to the old one:
* The returned object is different, in the old endpoint an array with
the users is returned. In the new endpoint an object of the type
`{"items": [...]}` is used.
* The returned users as with previous PRs is not including the
`workspaces` attribute and `full_name` has been removed as well.

Closes #141 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Adding new tests.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)