feat: migrate `POST /api/users` endpoint to `POST /api/v1/users` (#146)
# Description

This PR adds a new endpoint `POST /api/v1/users` allowing the creation
of a new user for API v1.

Changes of the new endpoint compared to the old one:
* `workspaces` attribute is not allowed when creating users with this
endpoint. If necessary subsequent requests should be done to an endpoint
that is still pending to be migrated. Specifically `POST
/api/v1/workspaces/:workspace_id/users/:user_id`.
* In the same way that is happening with previous PRs the returned
created user is not including `full_name` attribute.

Closes #145 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Adding new tests and checking that the tests for API v0 are
working as expected if using API v1 endpoint (except the ones creating
workspaces along with the user).

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)