feat: migrate `GET /api/workspaces/:workspace_id/users` endpoint to `GET /api/v1/workspaces/:workspace_id/users` (#153)
# Description

This PR adds a new endpoint `GET /api/v1/workspaces/:workspace_id/users`
allowing to get a list of users belonging to a workspace for API v1.

Changes of the new endpoint compared to the old one:
* The list of returned users are wrapped inside an `items` attribute
like `{"items": [...]}`. This is what we do for all endpoints on API v1
returning collections.
* The returned users are not including the workspaces. For the migration
we removed that attribute from `User` schema and that affect this
endpoint too.
* Small change on error message when the workspace is not found (we are
using a custom HTTP exception).

Closes #152 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

- [x] Adding new tests.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)