feat: migrate `POST /api/workspaces/:workspace_id/users` to API v1 (#156)
# Description

This PR add a new endpoint `POST /api/v1/workspaces/:workspace_id/users`
to create a new user for a workspace.

Changes introduced in this new endpoint compared to the old one:
* Now we must provide the `user_id` in the body of the `POST` request
instead of adding it in the URL.
* I have changed the `create_workspace_user` context function so we need
an additional change in `argilla` repo. The PR with the changes on
argilla is here: https://github.com/argilla-io/argilla/pull/4807
* As done with previous PRs the returned user doesn't has `full_name` or
`workspaces`.

Closes #155 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

- [x] Adding new tests for the endpoint.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)